+++ source/lucidity/template.py
@@ -2,13 +2,16 @@
 # :copyright: Copyright (c) 2013 Martin Pengelly-Phillips
 # :license: See LICENSE.txt.
 
+import os
+import re
+import imp
 import abc
 import sys
-import re
+import uuid
 import functools
 from collections import defaultdict
 
-import lucidity.error
+from . import error
 
 # Type of a RegexObject for isinstance check.
 _RegexType = type(re.compile(''))
@@ -42,7 +45,7 @@ class Template(object):
         be handled during parsing. :attr:`~Template.RELAXED` mode extracts the
         last matching value without checking the other values.
         :attr:`~Template.STRICT` mode ensures that all duplicate placeholders
-        extract the same value and raises :exc:`~lucidity.error.ParseError` if
+        extract the same value and raises :exc:`~error.ParseError` if
         they do not.
 
         If *template_resolver* is supplied, use it to resolve any template
@@ -84,8 +87,8 @@ class Template(object):
     def expanded_pattern(self):
         '''Return pattern with all referenced templates expanded recursively.
 
-        Raise :exc:`lucidity.error.ResolveError` if pattern contains a reference
-        that cannot be resolved by currently set template_resolver.
+        Raise :exc:`error.ResolveError` if pattern contains a
+        reference that cannot be resolved by currently set template_resolver.
 
         '''
         return self._TEMPLATE_REFERENCE_REGEX.sub(
@@ -97,14 +100,15 @@ class Template(object):
         reference = match.group('reference')
 
         if self.template_resolver is None:
-            raise lucidity.error.ResolveError(
-                'Failed to resolve reference {0!r} as no template resolver set.'
+            raise error.ResolveError(
+                'Failed to resolve reference {0!r} as no template '
+                'resolver set.'
                 .format(reference)
             )
 
         template = self.template_resolver.get(reference)
         if template is None:
-            raise lucidity.error.ResolveError(
+            raise error.ResolveError(
                 'Failed to resolve reference {0!r} using template resolver.'
                 .format(reference)
             )
@@ -114,7 +118,7 @@ class Template(object):
     def parse(self, path):
         '''Return dictionary of data extracted from *path* using this template.
 
-        Raise :py:class:`~lucidity.error.ParseError` if *path* is not
+        Raise :py:class:`~error.ParseError` if *path* is not
         parsable by this template.
 
         '''
@@ -131,12 +135,13 @@ class Template(object):
                 # Strip number that was added to make group name unique.
                 key = key[:-3]
 
-                # If strict mode enabled for duplicate placeholders, ensure that
-                # all duplicate placeholders extract the same value.
+                # If strict mode enabled for duplicate placeholders,
+                # ensure that all duplicate placeholders extract the
+                # same value.
                 if self.duplicate_placeholder_mode == self.STRICT:
                     if key in parsed:
                         if parsed[key] != value:
-                            raise lucidity.error.ParseError(
+                            raise error.ParseError(
                                 'Different extracted values for placeholder '
                                 '{0!r} detected. Values were {1!r} and {2!r}.'
                                 .format(key, parsed[key], value)
@@ -156,14 +161,14 @@ class Template(object):
             return data
 
         else:
-            raise lucidity.error.ParseError(
+            raise error.ParseError(
                 'Path {0!r} did not match template pattern.'.format(path)
             )
 
     def format(self, data):
         '''Return a path formatted by applying *data* to this template.
 
-        Raise :py:class:`~lucidity.error.FormatError` if *data* does not
+        Raise :py:class:`~error.FormatError` if *data* does not
         supply enough information to fill the template fields.
 
         '''
@@ -188,7 +193,7 @@ class Template(object):
                 value = value[part]
 
         except (TypeError, KeyError):
-            raise lucidity.error.FormatError(
+            raise error.FormatError(
                 'Could not format data {0!r} due to missing key {1!r}.'
                 .format(data, placeholder)
             )
@@ -208,7 +213,8 @@ class Template(object):
         format_specification = self._construct_format_specification(
             self.pattern
         )
-        return set(self._TEMPLATE_REFERENCE_REGEX.findall(format_specification))
+        return set(self._TEMPLATE_REFERENCE_REGEX.findall(
+            format_specification))
 
     def _construct_format_specification(self, pattern):
         '''Return format specification from *pattern*.'''
@@ -252,7 +258,7 @@ class Template(object):
             else:
                 _, value, traceback = sys.exc_info()
                 message = 'Invalid pattern: {0}'.format(value)
-                raise ValueError, message, traceback  #@IgnorePep8
+                raise ValueError(message, traceback)
 
         return compiled
 
@@ -324,3 +330,120 @@ class Resolver(object):
             return callable(getattr(subclass, 'get', None))
 
         return NotImplemented
+
+
+def discover_templates(paths=None, recursive=True):
+    '''Search *paths* for mount points and load templates from them.
+
+    *paths* should be a list of filesystem paths to search for mount points.
+    If not specified will try to use value from environment variable
+    :envvar:`LUCIDITY_TEMPLATE_PATH`.
+
+    A mount point is a Python file that defines a 'register' function. The
+    function should return a list of instantiated
+    :py:class:`~lucidity.template.Template` objects.
+
+    If *recursive* is True (the default) then all directories under a path
+    will also be searched.
+
+    '''
+    templates = []
+
+    if paths is None:
+        paths = os.environ.get('LUCIDITY_TEMPLATE_PATH', '').split(os.pathsep)
+
+    for path in paths:
+        for base, directories, filenames in os.walk(path):
+            for filename in filenames:
+                _, extension = os.path.splitext(filename)
+                if extension != '.py':
+                    continue
+
+                module_path = os.path.join(base, filename)
+                module_name = uuid.uuid4().hex
+                module = imp.load_source(module_name, module_path)
+                try:
+                    registered = module.register()
+                except AttributeError:
+                    pass
+                else:
+                    if registered:
+                        templates.extend(registered)
+
+            if not recursive:
+                del directories[:]
+
+    return templates
+
+
+def parse(path, templates):
+    '''Parse *path* against *templates*.
+
+    *path* should be a string to parse.
+
+    *templates* should be a list of :py:class:`~lucidity.template.Template`
+    instances in the order that they should be tried.
+
+    Return ``(data, template)`` from first successful parse.
+
+    Raise :py:class:`~error.ParseError` if *path* is not
+    parseable by any of the supplied *templates*.
+
+    '''
+    for template in templates:
+        try:
+            data = template.parse(path)
+        except error.ParseError:
+            continue
+        else:
+            return (data, template)
+
+    raise error.ParseError(
+        'Path {0!r} did not match any of the supplied template patterns.'
+        .format(path)
+    )
+
+
+def format(data, templates):  # @ReservedAssignment
+    '''Format *data* using *templates*.
+
+    *data* should be a dictionary of data to format into a path.
+
+    *templates* should be a list of :py:class:`~lucidity.template.Template`
+    instances in the order that they should be tried.
+
+    Return ``(path, template)`` from first successful format.
+
+    Raise :py:class:`~error.FormatError` if *data* is not
+    formattable by any of the supplied *templates*.
+
+
+    '''
+    for template in templates:
+        try:
+            path = template.format(data)
+        except error.FormatError:
+            continue
+        else:
+            return (path, template)
+
+    raise error.FormatError(
+        'Data {0!r} was not formattable by any of the supplied templates.'
+        .format(data)
+    )
+
+
+def get_template(name, templates):
+    '''Retrieve a template from *templates* by *name*.
+
+    Raise :py:exc:`~error.NotFound` if no matching template with
+    *name* found in *templates*.
+
+    '''
+    for template in templates:
+        if template.name == name:
+            return template
+
+    raise error.NotFound(
+        '{0} template not found in specified templates.'.format(name)
+    )
